rename ddbPut function to better represent function action and add ddbGet function that checks if the user phoneNumber is already in the db. If it is, it pulls the number directly.
